feat(annotation-thread): AnnotationThread component move to features
