+++ src/features/annotation-thread/AnnotationThread.scss
@@ -1,4 +1,4 @@
-@import '../common/variables';
+@import 'src/elements/common/variables';
 
 $annotationThreadWidth: 340px;
 $annotationThreadHeight: 600px;
